requirements.txt
apache-airflow==2.9
pandas==1.5.3
requests==2.28.1
beautifulsoup4==4.11.1
textblob==0.17.1
pytest==7.2.0
